shuffle tokenizer utils to a separate file, rename
